+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.17] - 2023-03-04
+
+### Fixed
+
+- Form didn't reset to initial values.
+- Reset also cleared last message.
+
 ## [0.5.16] - 2023-03-01
 
 ### Fixed